added child detail page
