+++ app/src/main/res/layout/activity_add_child.xml
@@ -5,76 +5,102 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ChildActivity">
+    tools:context=".AddChildActivity">
 
-    <include
-        android:id="@+id/appbar_id"
-        layout="@layout/app_bar" />
-
-    <TextView
+    <EditText
         android:id="@+id/input_fname"
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
-        android:layout_marginTop="25sp"
-        android:textAlignment="center"
-        android:textSize="32sp"
-        android:textStyle="bold"
+        android:inputType="text"
         android:hint="@string/hint_firstName"
         />
-    <TextView
-        android:id="@+id/input_isNaughty"
-        android:layout_width="match_parent"
+    <EditText
+        android:id="@+id/input_lname"
         android:layout_height="wrap_content"
-        android:layout_marginBottom="25sp"
-        android:textAlignment="center"
-        android:text="@string/hint_isNaughty"
+        android:layout_width="match_parent"
+        android:inputType="text"
+        android:hint="@string/hint_lastName"
         />
-    <TextView
+    <EditText
         android:id="@+id/input_bdate"
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
+        android:inputType="text"
         android:hint="@string/hint_bdate"
         />
-    <TextView
+    <EditText
         android:id="@+id/input_street"
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
+        android:inputType="text"
         android:hint="@string/hint_street"
         />
-    <TextView
+    <EditText
         android:id="@+id/input_city"
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
+        android:inputType="text"
         android:hint="@string/hint_city"
         />
-    <TextView
+    <EditText
         android:id="@+id/input_province"
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
+        android:inputType="text"
         android:hint="@string/hint_province"
         />
-    <TextView
+    <EditText
         android:id="@+id/input_postalCode"
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
+        android:inputType="text"
         android:hint="@string/hint_postalCode"
         />
-    <TextView
+    <EditText
         android:id="@+id/input_country"
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
+        android:inputType="text"
         android:hint="@string/hint_country"
         />
-    <TextView
+    <EditText
         android:id="@+id/input_lat"
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
+        android:inputType="text"
         android:hint="@string/hint_lat"
         />
-    <TextView
+    <EditText
         android:id="@+id/input_lng"
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
+        android:inputType="text"
         android:hint="@string/hint_lng"
         />
+
+    <CheckBox
+        android:id="@+id/input_isNaughty"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/hint_isNaughty"
+        android:checked="false" />
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        >
+        <Button
+            style="?android:attr/buttonBarButtonStyle"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/input_submit"
+            android:onClick="onClickSubmit"
+            />
+        <Button
+            style="?android:attr/buttonBarButtonStyle"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/input_cancel"
+            android:onClick="onClickCancel"/>
+    </LinearLayout>
 </LinearLayout>
